requirements.txt
asgiref==3.6.0
attrs==23.1.0
backports.zoneinfo==0.2.1
boto3==1.26.131
botocore==1.29.131
Django==4.1.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
importlib-resources==5.12.0
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.17.3
mysqlclient==2.1.1
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.15
zipp==3.15.0
